ability to edit/remove vidispine id. frontend will refuse to set an invalid ID and show a warning if id is invalid
